pyproject.toml
[tool.poetry]
name = "job-management-portal"
version = "0.1.0"
description = ""
authors = ["iana"]
readme = "README.md"
packages = [{include = "job_management_portal"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.3"
django-model-utils = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
